+++ NEWS.md
@@ -6,7 +6,9 @@
 
 ## Minor improvements and bug fixes
 
+* `as_pseudo_glycome()` now accepts a `GlycoproteomicSE` object and returns a `GlycomicSE` object. (#19)
 * `as_se()` now uses `"abundance"` instead of `"counts"` as the default`SummarizedExperiment` assay name. (#17)
+* `summarize_experiment()` now supports `GlycomicSE` and `GlycoproteomicSE` objects. (#19)
 
 # glyexp 0.14.2